pyproject.toml
[tool.poetry]
name = "test-task"
version = "0.1.0"
description = ""
authors = ["Pavel Ivanov <pavel.ivanov.060988@gmail.com>"]
readme = "README.md"
packages = [{include = "test_task"}]

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.23.1"
beautifulsoup4 = "^4.12.3"
django = "^5.0.7"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-phonenumber-field = "^8.0.0"
phonenumbers = "^8.13.40"
drf-spectacular = "^0.27.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
